Move ReportCardStep base into post_sim alongside Visualization
Co-locate the ReportCardStep base Step and REPORT_CARD_REGISTRY with the
Visualization base in workflow/post_sim.py (the unified post-sim registry
module), mirroring how the Analysis base lives in workflow/analysis.py —
outside its concrete-step package. report_cards/__init__.py now re-exports
both, so every consumer (concrete cards, flush, runner, tests) imports
unchanged.
