+++ v2ecoli/workflow/post_sim.py
@@ -12,6 +12,10 @@ KINDS = ("analysis", "visualization", "report_card")
 # name -> {"cls": <Step subclass>, "kind": <one of KINDS>}
 POST_SIM_REGISTRY: dict[str, dict] = {}
 
+# name -> ReportCardStep subclass. Populated by __init_subclass__ for any
+# subclass that defines its own ``name`` (mirrors ANALYSIS_REGISTRY).
+REPORT_CARD_REGISTRY: dict[str, type] = {}
+
 
 def register_post_sim(cls, kind: str, name: "str | None" = None) -> None:
     """Register a post-sim Step subclass under ``name`` (default ``cls.name``)
@@ -63,3 +67,43 @@ class Visualization(V2Step):
             return {"view": "", "data": {}}
         view, data = res
         return {"view": view, "data": data}
+
+
+class ReportCardStep(V2Step):
+    """A report card as a visualization-like Step (sibling of ``Analysis`` and
+    ``Visualization``): emits ``view`` (HTML) + ``data`` (verdict map). Unlike
+    ``Analysis`` — which consumes a live DuckDB sim-output connection — a report
+    card's input is a ``StudyContext`` (the study's spec + dir), so cards grade
+    run-free. Subclasses set ``name`` and implement ``applies(study)`` +
+    ``build(study) -> (verdict_dict, html) | None``. A named subclass auto-registers
+    in ``REPORT_CARD_REGISTRY`` and, kind-tagged, in ``POST_SIM_REGISTRY``."""
+
+    name: str = ""
+    config_schema: dict = {}
+
+    def __init_subclass__(cls, **kwargs):
+        super().__init_subclass__(**kwargs)
+        if cls.__dict__.get("name"):
+            REPORT_CARD_REGISTRY[cls.name] = cls
+            register_post_sim(cls, "report_card")
+
+    def inputs(self):
+        return {"study": "any"}
+
+    def outputs(self):
+        return {"view": "string", "data": "map"}
+
+    def applies(self, study) -> bool:
+        return True
+
+    def build(self, study) -> "tuple[dict, str] | None":
+        """Return ``(verdict_json_dict, html_str)`` or None. Subclasses override."""
+        raise NotImplementedError
+
+    def update(self, state, interval=None):
+        study = state.get("study")
+        res = self.build(study) if study is not None else None
+        if not res:
+            return {"view": "", "data": {}}
+        verdict, html = res
+        return {"view": html, "data": verdict}
